bug fix
user: create, update + middleware
comment: security

+++ controllers/comment.controller.ts
@@ -1,17 +1,9 @@
 import express from 'express';
 import CommentRepository from "../models/comment.repository";
+import PostRepository from "../models/post.repository";
 
 const router = express.Router();
 
-router.get('/count/:id', async (req, res) => {
-    const post_id = Number(req.params.id);
-    try {
-        const comments = await CommentRepository.getAllCommentCountById(post_id);
-        return res.status(200).json(comments[0]).end();
-    } catch (error: any) {
-        return res.status(500).json({message: error.message}).end();
-    }
-});
 
 router.get('/all/:id', async (req, res) => {
     const currentLoad = Number(req.query.currentLoad) || 1;
@@ -31,8 +23,9 @@ router.get('/all/:id', async (req, res) => {
 
 router.post('/create/:id', async (req, res) => {
     try {
-        const comment = await CommentRepository.createComment(Number(req.params.id), req.body.sentence, Number(req.body.user.id));
-        res.status(200).json({message: 'Comment created successfully!'}).end();
+        const created_comment = await CommentRepository.createComment(Number(req.params.id), req.body.sentence, Number(req.body.user.id));
+        const [find] = await CommentRepository.findComment(created_comment.insertId);
+        res.status(200).json(find).end();
     } catch (error: any) {
         res.status(500).json({message: error.message}).end();
     }
@@ -40,8 +33,14 @@ router.post('/create/:id', async (req, res) => {
 
 router.put('/update/:id', async (req, res) => {
     try {
-        const comment = await CommentRepository.updateComment(Number(req.params.id), req.body.sentence);
-        res.status(200).json({message: 'Comment updated successfully!'}).end();
+        const [find] = await CommentRepository.findComment(Number(req.params.id));
+        if (find.user_id === req.body.user.id) {
+            const comment = await CommentRepository.updateComment(Number(req.params.id), req.body.sentence);
+            res.status(200).json({message: 'Comment updated successfully!'}).end();
+        } else {
+            return res.status(403).json({message: 'Bad permissions!'}).end();
+        }
+
     } catch (error: any) {
         res.status(500).json({message: error.message}).end();
     }
@@ -49,8 +48,14 @@ router.put('/update/:id', async (req, res) => {
 
 router.delete('/delete/:id', async (req, res) => {
     try {
-        const comment = await CommentRepository.deleteComment(Number(req.params.id));
-        res.status(200).json({message: 'Comment deleted successfully!'}).end();
+        const [find] = await CommentRepository.findComment(Number(req.params.id));
+        if (find.user_id === req.body.user.id) {
+            const comment = await CommentRepository.deleteComment(Number(req.params.id));
+            res.status(200).json({message: 'Comment deleted successfully!'}).end();
+        } else {
+            return res.status(403).json({message: 'Bad permissions!'}).end();
+        }
+
     } catch (error: any) {
         res.status(500).json({message: error.message}).end();
     }
